[Refactor] #16 - SchoolBoxView 파일 분리
